Extract breadcrumbs to slot plugin
